+++ lib/langchain/llm/hugging_face.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 module Langchain::LLM
+  #
+  # Wrapper around the HuggingFace Inference API: https://huggingface.co/inference-api
+  #
+  # Gem requirements:
+  #     gem "hugging-face", "~> 0.3.4"
+  #
+  # Usage:
+  #     hf = Langchain::LLM::HuggingFace.new(api_key: "YOUR_API_KEY")
+  #
   class HuggingFace < Base
-    #
-    # Wrapper around the HuggingFace Inference API.
-    #
-    # Gem requirements: gem "hugging-face", "~> 0.3.4"
-    #
-    # Usage:
-    # hf = Langchain::LLM::HuggingFace.new(api_key: "YOUR_API_KEY")
-    #
-
     # The gem does not currently accept other models:
     # https://github.com/alchaplinsky/hugging-face/blob/main/lib/hugging_face/inference_api.rb#L32-L34
     DEFAULTS = {