+++ spec/langchain/agent/sql_query_agent/sql_query_agent_spec.rb
@@ -28,8 +28,7 @@ RSpec.describe Langchain::Agent::SQLQueryAgent do
 
     before do
       allow(subject.llm).to receive(:complete).with(
-        prompt: original_prompt,
-        max_tokens: 500
+        prompt: original_prompt
       ).and_return(llm_first_response)
 
       allow(Langchain::Tool::Database).to receive(:execute).with(
@@ -37,8 +36,7 @@ RSpec.describe Langchain::Agent::SQLQueryAgent do
       ).and_return(database_tool_response)
 
       allow(subject.llm).to receive(:complete).with(
-        prompt: "Given an input question and results of a SQL query, look at the results and return the answer. Use the following format:\nQuestion: What is the longest length name in the users table?\nThe SQL query: SQLQuery: SELECT name, LENGTH(name) FROM users HAVING MAX(length);\nResult of the SQLQuery: []\nFinal answer: Final answer here",
-        max_tokens: 500
+        prompt: "Given an input question and results of a SQL query, look at the results and return the answer. Use the following format:\nQuestion: What is the longest length name in the users table?\nThe SQL query: SQLQuery: SELECT name, LENGTH(name) FROM users HAVING MAX(length);\nResult of the SQLQuery: []\nFinal answer: Final answer here"
       ).and_return(llm_final_response)
     end